Fixes for non existing resources during clean and mark as derivated.
git-svn-id: https://svn.objectstyle.org/repos/woproject/trunk@2591 930f2ab1-3212-0410-a8ea-d1eb4cc50a93

+++ wolips/plugins/org.objectstyle.wolips/java/org/objectstyle/wolips/core/resources/internal/types/folder/BuildAdapter.java
@@ -102,6 +102,9 @@ public class BuildAdapter extends AbstractFolderAdapter implements
 	}
 
 	public void clean(IProgressMonitor monitor) {
+		if(!this.getUnderlyingFolder().exists()) {
+			return;
+		}
 		try {
 			this.getUnderlyingFolder().delete(true, monitor);
 		} catch (CoreException e) {
